Changed test to use delegate
http://jira.diamond.ac.uk/browse/DAQ-357
