+++ org.eclipse.scanning.test/src/org/eclipse/scanning/test/BrokerTest.java
@@ -33,21 +33,14 @@ import uk.ac.diamond.daq.activemq.connector.ActivemqConnectorService;
 public class BrokerTest extends TmpTest {
 
 	protected static URI uri;     
-	
-	private static BrokerService service;
+
+	private static BrokerDelegate delegate;
 
 	@BeforeClass
 	public final static void startBroker() throws Exception {
-		uri = createUri(); // Each test uses a new port if the port is running on another test.
-		System.setProperty("org.eclipse.scanning.broker.uri", uri.toString());
-        service = new BrokerService();
-        service.addConnector(uri);
-        service.setPersistent(false); 
-        SystemUsage systemUsage = service.getSystemUsage();
-        systemUsage.getStoreUsage().setLimit(1024 * 1024 * 8);
-        systemUsage.getTempUsage().setLimit(1024 * 1024 * 8);
-        service.start();
-		service.waitUntilStarted();
+		delegate = new BrokerDelegate();
+		delegate.start();
+		uri      = delegate.getUri();
 	}
 	
 	public final static void setUpNonOSGIActivemqMarshaller() {
@@ -62,66 +55,9 @@ public class BrokerTest extends TmpTest {
 
 	@AfterClass
 	public final static void stopBroker() throws Exception {
-		
-		service.stop();
-		service.waitUntilStopped();
-		service = null;
+		delegate.stop();
 	}
 
-	private static URI createUri() {
-		try {
-			return new URI("tcp://localhost:"+getFreePort());
-		} catch (Exception ne) {
-			ne.printStackTrace();
-		    return null;
-		}
-	}
-
-	private static int getFreePort() {
-		final int start = 8619+Math.round((float)Math.random()*100);
-		return getFreePort(start);
-	}
-
-	
-	private static int getFreePort(final int startPort) {
-
-	    int port = startPort;
-	    while(!isPortFree(port)) port++;
-
-	    return port;
-	}
-	/**
-	 * Checks if a port is free.
-	 * @param port
-	 * @return
-	 */
-	public static boolean isPortFree(int port) {
-
-	    ServerSocket ss = null;
-	    DatagramSocket ds = null;
-	    try {
-	        ss = new ServerSocket(port);
-	        ss.setReuseAddress(true);
-	        ds = new DatagramSocket(port);
-	        ds.setReuseAddress(true);
-	        return true;
-	    } catch (IOException e) {
-	    } finally {
-	        if (ds != null) {
-	            ds.close();
-	        }
-
-	        if (ss != null) {
-	            try {
-	                ss.close();
-	            } catch (IOException e) {
-	                /* should not be thrown */
-	            }
-	        }
-	    }
-
-	    return false;
-	}
 
 
 }
